requirements.txt
faiss-cpu
sentence-transformers
numpy
fastapi
uvicorn
pydantic
json
